dom/media/mediasource/test/test_AVC3_mp4.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: AVC3 content playback.</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(function(ms, el) {

  once(ms, 'sourceopen').then(function() {
    ok(true, "Receive a sourceopen event");
    var videosb = ms.addSourceBuffer("video/mp4");

    fetchAndLoad(videosb, 'avc3/init', [''], '.mp4')
    .then(fetchAndLoad.bind(null, videosb, 'avc3/segment', range(1, 2), '.m4s'))
    .then(function() {
      is(videosb.buffered.length, 1, "continuous buffered range");
      return once(el, "loadeddata");
    }).then(function() {
      ok(true, "got loadeddata");
      ms.endOfStream();
      return once(ms, "sourceended");
    }).then(function() {
      ok(true, "endOfStream completed");
      // Now ensure that we can play to the end.
      el.play();
      once(el, 'ended').then(SimpleTest.finish.bind(SimpleTest));
    });
  });
});

</script>
</pre>
</body>
</html>
